Add TextSelectionManager as NSTextStorage delegate
